Feature/uwp nuget 2

See merge request BC/public/mediastreamer2!355

+++ src/utils/mkv_reader.cpp
@@ -17,12 +17,14 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <stdint.h>
 #include <algorithm>
 #include <array>
 #include <cwchar>
 #include <list>
 #include <memory>
 #include <vector>
+
 
 #define bool_t mkv_bool_t
 extern "C" {
@@ -65,7 +67,7 @@ static std::string mkvStringToCppString(const MkvStrT &mkvString) {
 #ifdef UNICODE
 	strOut.resize(wcslen(mkvString.data()) * 2);
 #ifdef _WIN32
-	WideCharToMultiByte(CP_UTF8, 0, mkvString.data(), -1, &strOut[0], mkvString.size(), NULL, NULL);
+	WideCharToMultiByte(CP_UTF8, 0, mkvString.data(), -1, &strOut[0], (int)mkvString.size(), NULL, NULL);
 #else
 	wcstombs(&strOut[0], mkvString.data(), strOut.size());
 #endif
